Add code splitting for loading indicator
